Move gcm_climate to oggm.shop
